+++ managers/parsers/__init__.py
@@ -4,3 +4,4 @@ from .frontierParser import FrontierParser
 from .mdpiParser import MDPIParser
 from .openEditionParser import OpenEditionParser
 from .springerParser import SpringerParser
+from .inTechOpenParser import InTechOpenParser