+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Work title and author info to edition endpoint
 ### Fixed
 - Work title and author info only shows up in edition endpoint now
+- Editions and utils tests updated
 
 ## 2022-05-19 -- v0.10.3
 ### Added