+++ CHANGELOG.md
@@ -4,12 +4,16 @@
 ### Added
 - Updated format/filter mappings in APIUtils class
 - Work title and author info to edition endpoint
+- Added intechopen parser 
+- Updated tests for DOAB Link Manager
 ### Fixed
-- Work title and authors info only show up in edition endpoint now
+- Work title and authors info only show up in edition details endpoint now
 - Editions and utils tests updated and fixed
 - All toc object url links in ProjectMuse manifests have a query parameter now
 - Update DOAB single record import to reflect new OAI-PMH structure
 - Deleted extra urls in toc object of Muse manifests 
+- Read Online links for books from IntechOpen website
+
 
 ## 2022-05-19 -- v0.10.3
 ### Added